cicd: everything should work now
