+++ .github/workflows/test.yml
@@ -9,8 +9,7 @@ jobs:
     steps:
       - name: Get current date
         id: date
-        run: |
-          echo "$(date +'%Y-%m-%d-%s %T')" > date.txt
+        run: echo "$(date +'%Y-%m-%d-%s %T')" > date.txt
 
       - uses: actions/checkout@v2
 
@@ -50,9 +49,16 @@ jobs:
     runs-on: self-hosted
     steps:
     - name: Discord Notification
+      if: ${{ success() }}
       env:
         DISCORD_BOT_TOKEN: ${{ secrets.DISCORD_BOT_TOKEN }}
       run: |
         pip3 install -r requirements.txt
-        python3 discord_bot.py -t "$(<date.txt)" -b ${{github.ref_name}} -s ${{ github.event.workflow_run.conclusion == 'success' }} -p ${{ github.job }}
-        rm date.txt
+        python3 .github/discord_python/discord_bot.py -t "$(date +'%Y-%m-%d-%s %T')" -b ${{ github.ref_name }} -s 'success' -p ${{ github.job }} -a ${{ github.actor }}
+    - name: Discord Notification
+      if: ${{ failure() }}
+      env:
+        DISCORD_BOT_TOKEN: ${{ secrets.DISCORD_BOT_TOKEN }}
+      run: |
+        pip3 install -r requirements.txt
+        python3 .github/discord_python/discord_bot.py -t "$(date +'%Y-%m-%d-%s %T')" -b ${{ github.ref_name }} -s 'fail' -p ${{ runner.name }} -a ${{ github.actor }}
